Created Insight360 App Sub Module
This was done so that that the whole i360 Module would consist of a loading page and the main app feature module. The router could then display the loading component until the lazy loading of the main app feature module is completed. Using ngIf is a possibility with this but it wouldn't give the desired effect.